Add files via upload
Solved the following questions:
1. Write a Python program to remove duplicate elements from a tuple.
2. Write a Python program to form a set of elements present in set A or set B, but not common
to both the sets.
3. Write a Python program to create a dictionary of phone numbers and names of five persons.
Display the contents of the dictionary in alphabetical order of names.
4. A book shop details contains the Title of the book and number of copies of each title. As
books are added to the shop, the number of copies in each should increase and as books are
sold, the number of copies in each should decrease. Implement this using dictionary data
type.
5. Write a Python program to create a dictionary with details(roll no, name, mark1, mark2,
mark3) of 5 students. Compute the total marks for each student and display the details. Also
find the student with highest total.